+++ packages/datadog_flutter_plugin/CHANGELOG.md
@@ -6,6 +6,7 @@
 * Fix default console print functions. See [#575] and [#574]
 * Loosen the restriction on the `js` package to allow `<0.8`. See [#572]
 * Add support for global attributes for logs.
+* Passing `null` to `addAttribute` now calls `removeAttribute` instead of silently failing.
 
 ## 2.3.0